requirements.txt
flask
pymongo
bcrypt
cryptography
argon2-cffi
python-dotenv
requests